create component and template sol price
